requirements/base.txt
# Versions are left out on purpose in order for you to use the latest ones.
# After the first "pip install -r", just run "pip freeze" and add the version
# to each package in each requirements/*.txt.

Django>=1.10,<1.11

django-authtools>=1.5
django-extensions>=1.7
django-model-utils>=2.6
django-choices>=1.4

psycopg2>=2.6

channels>=1.0.2
asgi_redis>=1.0.0

social-auth-app-django>=1.0.1
